requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
apache-libcloud==2.5.0
    # via salt-ssh
attrs==21.2.0
    # via pytest
build==0.9.0
    # via pip-tools
certifi==2022.5.18.1
    # via
    #   elasticsearch
    #   requests
    #   salt-ssh
cffi==1.14.6
    # via
    #   cryptography
    #   salt-ssh
chardet==3.0.4
    # via
    #   requests
    #   salt-ssh
cheroot==8.5.2
    # via
    #   cherrypy
    #   salt-ssh
cherrypy==18.6.1
    # via salt-ssh
click==8.0.1
    # via
    #   -r requirements.in
    #   ocdsindex
    #   pip-tools
contextvars==2.4
    # via salt-ssh
cryptography==3.4.6
    # via
    #   pyopenssl
    #   salt-ssh
distro==1.5.0
    # via salt-ssh
elasticsearch==7.13.1
    # via ocdsindex
gitdb==4.0.5
    # via
    #   gitpython
    #   salt-ssh
gitpython==3.1.12
    # via salt-ssh
idna==2.8
    # via
    #   requests
    #   salt-ssh
immutables==0.15
    # via
    #   contextvars
    #   salt-ssh
iniconfig==1.1.1
    # via pytest
jaraco-classes==3.2.1
    # via
    #   jaraco-collections
    #   salt-ssh
jaraco-collections==3.4.0
    # via
    #   cherrypy
    #   salt-ssh
jaraco-functools==2.0
    # via
    #   cheroot
    #   jaraco-text
    #   salt-ssh
    #   tempora
jaraco-text==3.5.1
    # via
    #   jaraco-collections
    #   salt-ssh
jinja2==3.1.0
    # via salt-ssh
jmespath==1.0.1
    # via salt-ssh
linode-python==1.1.1
    # via salt-ssh
lxml==4.9.1
    # via ocdsindex
mako==1.1.4
    # via salt-ssh
markupsafe==2.0.1
    # via
    #   jinja2
    #   mako
    #   salt-ssh
more-itertools==8.2.0
    # via
    #   cheroot
    #   cherrypy
    #   jaraco-classes
    #   jaraco-functools
    #   salt-ssh
msgpack==1.0.2
    # via salt-ssh
ocdsindex==0.0.7
    # via -r requirements.in
packaging==20.9
    # via
    #   build
    #   pytest
pep517==0.10.0
    # via build
pip-tools==6.11.0
    # via -r requirements.in
pluggy==0.13.1
    # via pytest
portend==2.6
    # via
    #   cherrypy
    #   salt-ssh
psutil==5.8.0
    # via salt-ssh
py==1.10.0
    # via pytest
pyasn1==0.4.8
    # via salt-ssh
pycparser==2.21
    # via
    #   cffi
    #   salt-ssh
pycryptodomex==3.9.8
    # via salt-ssh
pyopenssl==19.0.0
    # via salt-ssh
pyparsing==2.4.7
    # via packaging
pytest==6.2.4
    # via -r requirements.in
python-dateutil==2.8.0
    # via salt-ssh
python-gnupg==0.4.8
    # via salt-ssh
pytz==2022.1
    # via
    #   salt-ssh
    #   tempora
pyyaml==5.4.1
    # via salt-ssh
pyzmq==23.2.0
    # via salt-ssh
requests==2.25.1
    # via
    #   -r requirements.in
    #   apache-libcloud
    #   salt-ssh
    #   vultr
salt-ssh==3005.1
    # via -r requirements.in
setproctitle==1.3.2
    # via salt-ssh
six==1.16.0
    # via
    #   cheroot
    #   cryptography
    #   pyopenssl
    #   python-dateutil
    #   salt-ssh
smmap==3.0.2
    # via
    #   gitdb
    #   salt-ssh
tempora==4.1.1
    # via
    #   portend
    #   salt-ssh
timelib==0.2.5
    # via salt-ssh
toml==0.10.2
    # via pytest
tomli==2.0.1
    # via build
urllib3==1.26.6
    # via
    #   elasticsearch
    #   requests
    #   salt-ssh
vultr==1.0.1
    # via salt-ssh
wheel==0.38.4
    # via pip-tools
zc-lockfile==2.0
    # via
    #   cherrypy
    #   salt-ssh

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
